Website Navigation Setup
Instructions if lost are in app.js
