+++ memomate/package.json
@@ -10,6 +10,7 @@
     "postcss": "^8.4.31",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.19.0",
     "react-scripts": "5.0.1",
     "tailwindcss": "^3.3.5",
     "web-vitals": "^2.1.4"
